Moved the sitter/flags.cpp/h to communicatord.
The sitter depends on communicatord (which is fixed now) and thus we can't
have communicatord depend on the sitter. So I moved the flags even though it
makes more sense in the sitter but without that the communicatord could not
itself raise flags. Later we may even move that class further down...