Nearly working version of the sitter.
We still have an issue with the JSON, it looks like. But this version runs well.
Updated the Messenger so it derives from fluid-settings (and thus communicatord).
Properly updated the .ini files with group=options & allowed=dynamic-configuration.
Moved the "disk-ignore=..." parameter to the sitter_disk plugin.
Make use of the duration validator to read many of the durations (at least in the main server).
Actually added many of the depdencies to the sitter target.
Added .ini files for all the plugins that have dynamic parameters (read using fluid).
Enable the ticks only once we received the fluid-ready message (i.e. once we know all the parameters are properly defined).
Fixed a couple of variables in the sitter processes names.an file.
Made the f_scripts variable initialization work (we need a real object with a position and the name must not be empty).
Broke the sitter.cpp file in 3 more files (interrupt, messenger, tick_timer). Each is much smaller now.
Replaced the hard coded command line parameters with the dynamically loaded fluid-settings.
Changed the configuration filename to make use of the f_configuration_filename instead of the f_configuration_files (i.e. we get the list of directories defined automatically).
Updated the plugins to match the new serverplugins initialization scheme.
Removed the log rotate special implementation. We are connected to the communicator daemon which manes it's 100% handled automatically.
Made the connections member variables instead of globals since I moved them to separate files.
Initialize parameters in a way allowing for fluid-settings updates.
Added parameter min/max/default as constexpr instead of being hard coded.
Added the "sitter_" prefix name to plugins that do not have it.
Changed many fatal errors to recoverable errors when loading parameters from the fluid-settings so we can instead use the default, minimum or maximum.
Commented out the CASSANDRASTATUS message. At this time, this is not ready. Also, the snapdatabase can be responsible for doing so on its own.