Made sure we detect when the thread dies.
If the thread gets an exception or exits, we want to know about it. For this
I have now a worker_done connection which receives a signal from the thread.
This allows us to stop the whole process because if the thread dies, the
sitter is finished.

Also, I added the necessary to stop the thread from the stop() function. It
is not mandatory, but cleaner this way. If the thread is currently processing,
through, it will go on until done (the boost signal does not offer us means
to exit early).