Broke the run() function in 3.
The cppthread offers the enter(), run(), and leave() functions.
I now use enter() to do the initliazation (loading of plugins).
run() will run the loop().
And the leave() makes sure that the process quits (i.e. if an error occurs
and the thread dies, we want to quit the entire service because at that point
it is officially not working anymore).