Got a first version that loops as expected.
Removed many e["error"] = ... in plugins. Instead we use the append_error() function.
Changed the partition "full" error to include "(ignore)" when the patterns match the path.
Fixed the arrays (although we still have some issues with as2js where we can't add a new item at the end of the array if we don't have the parent object available).
Switched the communicator to use timespec_ex for without extra precision, the test on whether it runs fails.
Fixed the exiting of the service, we have to properly stop the thread by setting f_ticks to 1.
Also when exiting we do not want to run the plugins.
The "/proc/loadavg" was set to "/proc/uptime".
Cleaned up many files that still had many unused #include.
Added a TODO.